New database integrated into docker. Queries for homepage rewritten to work with new DB. Old/outdated sql files stashed in deadsql folder (for now). Added a notes text file with instructions to connect to the DB from within docker
