+++ _deadsql/postgres/csv/discounts.csv
@@ -1,0 +1,1 @@
+id,code,amount,enddate,startdate,usagelimit,min_tickets,min_events
