+++ _deadsql/postgres/csv/event_instances.csv
@@ -1,0 +1,33 @@
+id,eventid,eventdate,starttime,salestatus,totalseats,availableseats,purchaseuri
+1,1,2021-01-07,19:00:00,t,67,2,https://portlandplayhouse.org/oa17T
+8,1,2021-01-14,19:00:00,t,87,5,https://portlandplayhouse.org/1DPj5L
+15,1,2021-01-21,19:00:00,t,90,13,https://portlandplayhouse.org/3gQ6Dh
+22,1,2021-01-28,19:00:00,t,79,16,https://portlandplayhouse.org/27tkuc
+29,1,2021-02-07,19:00:00,t,68,6,https://portlandplayhouse.org/3UVqwX
+34,2,2021-02-27,19:00:00,t,75,3,https://portlandplayhouse.org/3z.g_F
+41,2,2021-03-06,19:00:00,t,79,12,https://portlandplayhouse.org/1ccIrB
+48,2,2021-03-13,19:00:00,t,79,2,https://portlandplayhouse.org/1F7gdc
+55,2,2021-03-20,19:00:00,t,70,8,https://portlandplayhouse.org/1uajTz
+62,2,2021-03-27,19:00:00,t,85,4,https://portlandplayhouse.org/2CVvoR
+67,3,2021-04-19,19:00:00,t,88,16,https://portlandplayhouse.org/1Z7Kpo
+74,3,2021-04-26,19:00:00,t,76,14,https://portlandplayhouse.org/2mjo9d
+81,3,2021-05-05,19:00:00,t,80,8,https://portlandplayhouse.org/3FlXr5
+89,3,2021-05-13,19:00:00,t,72,12,https://portlandplayhouse.org/3HlXZQ
+96,3,2021-05-20,19:00:00,t,67,2,https://portlandplayhouse.org/2DXU1Z
+100,4,2021-06-11,19:00:00,t,79,13,https://portlandplayhouse.org/xSyqi
+107,4,2021-06-18,19:00:00,t,90,11,https://portlandplayhouse.org/27PRVd
+114,4,2021-06-25,19:00:00,t,62,4,https://portlandplayhouse.org/2Dm5Yl
+121,4,2021-07-04,19:00:00,t,89,3,https://portlandplayhouse.org/2snmM6
+129,4,2021-07-12,19:00:00,t,75,3,https://portlandplayhouse.org/ZNgDu
+133,5,2021-08-03,19:00:00,f,86,0,https://portlandplayhouse.org/kXqqd
+140,5,2021-08-10,19:00:00,t,69,7,https://portlandplayhouse.org/1oq2t5
+147,5,2021-08-17,19:00:00,t,74,9,https://portlandplayhouse.org/ndEV.
+154,5,2021-08-24,19:00:00,t,68,14,https://portlandplayhouse.org/3H1Dea
+161,5,2021-09-03,19:00:00,t,83,7,https://portlandplayhouse.org/3fHEPQ
+166,6,2021-09-23,19:00:00,t,76,16,https://portlandplayhouse.org/3r8WX3
+173,6,2021-10-02,19:00:00,f,88,0,https://portlandplayhouse.org/1g.F7w
+180,6,2021-10-09,19:00:00,t,86,13,https://portlandplayhouse.org/2NNaPl
+187,6,2021-10-16,19:00:00,t,63,8,https://portlandplayhouse.org/2jkdSb
+194,6,2021-10-23,19:00:00,t,77,7,https://portlandplayhouse.org/2U_7KQ
+200,8,2021-11-27,22:00:00,t,100,100,
+201,8,2021-12-25,22:00:00,t,100,100,
