+++ _deadsql/postgres/csv/events.csv
@@ -1,0 +1,8 @@
+id,seasonid,eventname,eventdescription,active,image_url
+2,0,staid of Gheen,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://www.artswest.org/wp-content/uploads/2013/08/Godspell-2-615x461.jpg
+1,0,The last cafe,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://www.kansas.com/latest-news/nnxqa/picture251039064/alternates/FREE_1140/601A1029.jpg
+4,0,Brianhead of egg,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://playstosee.com/wp-content/uploads/2016/03/Egg-Theodora-Van-der-Beek-Camden-Peoples-Theatre-01.jpe
+6,0,halt of scholar,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://s7d2.scene7.com/is/image/TWCNews/dukeanderson
+5,0,perfection of grasp,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://simonparrismaninchair.files.wordpress.com/2019/01/the-inheritance-west-end-samuel-h-levine-kyle-soller-andrew-burnap.jpg
+3,0,united,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Aenean et tortor at risus viverra. Arcu non odio euismod lacinia. Faucibus scelerisque eleifend donec pretium vulputate sapien nec",t,https://live.staticflickr.com/113/285441219_05b3265345_b.jpg
+8,0,test,test,t,https://www.pexels.com/photo/scientific-calculator-5775/
