+++ _deadsql/postgres/csv/exdoorlist.csv
@@ -1,0 +1,49 @@
+id,name,vip,donor,accomodations,num_tickets,arrived
+1,John Doe,t,t,t,1,f
+2,Sarah Doe,f,f,t,2,f
+3,Jane Doe,f,t,t,3,f
+4,Mike Doe,f,t,t,3,f
+5,Allan Doe,t,f,f,4,f
+6,Bob Doe,f,t,t,3,f
+7,Jon Doe,f,f,t,2,f
+8,Andrew Doe,t,t,t,1,f
+9,Barbara Doe,f,t,f,2,f
+10,Sherri Doe,f,t,t,1,f
+11,Wendy Doe,t,t,t,4,f
+12,Abe Doe,f,t,t,3,f
+1,John Doe,t,t,t,1,f
+2,Sarah Doe,f,f,t,2,f
+3,Jane Doe,f,t,t,3,f
+4,Mike Doe,f,t,t,3,f
+5,Allan Doe,t,f,f,4,f
+6,Bob Doe,f,t,t,3,f
+7,Jon Doe,f,f,t,2,f
+8,Andrew Doe,t,t,t,1,f
+9,Barbara Doe,f,t,f,2,f
+10,Sherri Doe,f,t,t,1,f
+11,Wendy Doe,t,t,t,4,f
+12,Abe Doe,f,t,t,3,f
+1,John Doe,t,t,t,1,f
+2,Sarah Doe,f,f,t,2,f
+3,Jane Doe,f,t,t,3,f
+4,Mike Doe,f,t,t,3,f
+5,Allan Doe,t,f,f,4,f
+6,Bob Doe,f,t,t,3,f
+7,Jon Doe,f,f,t,2,f
+8,Andrew Doe,t,t,t,1,f
+9,Barbara Doe,f,t,f,2,f
+10,Sherri Doe,f,t,t,1,f
+11,Wendy Doe,t,t,t,4,f
+12,Abe Doe,f,t,t,3,f
+1,John Doe,t,t,t,1,f
+2,Sarah Doe,f,f,t,2,f
+3,Jane Doe,f,t,t,3,f
+4,Mike Doe,f,t,t,3,f
+5,Allan Doe,t,f,f,4,f
+6,Bob Doe,f,t,t,3,f
+7,Jon Doe,f,f,t,2,f
+8,Andrew Doe,t,t,t,1,f
+9,Barbara Doe,f,t,f,2,f
+10,Sherri Doe,f,t,t,1,f
+11,Wendy Doe,t,t,t,4,f
+12,Abe Doe,f,t,t,3,f
