+++ _deadsql/postgres/csv/seasons.csv
@@ -1,0 +1,2 @@
+id,name,startdate,enddate
+0,Test Season Name,2021-01-07 00:00:00,2021-10-27 00:00:00
