+++ _deadsql/postgres/csv/tickettype.csv
@@ -1,0 +1,2 @@
+id,name,isseason,seasonid,price,concessions
+0,General Admission,f,0,$15.00,$5.00
