+++ _deadsql/postgres/csv/users.csv
@@ -1,0 +1,6 @@
+id,username,pass_hash,is_superadmin
+4,wtix,$2a$10$tRaShlhaRUDQ1ADT242dN.x2DSxstwHeHdc9hejnoV4doEbTX5Qfm,t
+0,brian,"",f
+1,hailey,"",f
+2,aiyana,"",f
+3,allen,"",f
