--- postgres/csv/discounts.csv
@@ -1,1 +1,0 @@
-id,code,amount,enddate,startdate,usagelimit,min_tickets,min_events
