--- postgres/csv/linkedtickets.csv
@@ -1,33 +1,0 @@
-id,event_instance_id,ticket_type
-4,1,0
-11,8,0
-18,15,0
-25,22,0
-32,29,0
-37,34,0
-44,41,0
-51,48,0
-58,55,0
-65,62,0
-70,67,0
-77,74,0
-84,81,0
-92,89,0
-99,96,0
-103,100,0
-110,107,0
-117,114,0
-124,121,0
-132,129,0
-136,133,0
-143,140,0
-150,147,0
-157,154,0
-164,161,0
-169,166,0
-176,173,0
-183,180,0
-190,187,0
-197,194,0
-202,200,0
-203,201,0
