--- postgres/csv/tickettype.csv
@@ -1,2 +1,0 @@
-id,name,isseason,seasonid,price,concessions
-0,General Admission,f,0,$15.00,$5.00
