--- sql-scripts/queries_todo.md
@@ -1,13 +1,0 @@
-# Queries To Do
-
-This folder contains two other files, one containing SQL code to create the tables, and one with completed SQL queries for possible reports.  
-This document will contain a list of possible queries to create SQL code for in no particular order.
-
-**Queries:**
-
-- [x] Daily Donors
-- [x] Door list
-- [x] Tickets sold by show
-- [x] Donation Threshold
-- [x] People who signs up for newsletter
-- [ ] Top XX Donors
